+++ Examples/utilities/timeout.robot
@@ -1,5 +1,4 @@
 *** Settings ***
-Force Tags    Ignore
 Library    PuppeteerLibrary
 Test Setup    Open browser to test page
 Test Teardown    Close Browser
