+++ PuppeteerLibrary/keywords/browsermanagement.py
@@ -1,4 +1,3 @@
-from robot.utils import timestr_to_secs
 from PuppeteerLibrary.base.librarycomponent import LibraryComponent
 from PuppeteerLibrary.base.robotlibcore import keyword
 from PuppeteerLibrary.keywords.browsermanagement_async import BrowserManagementKeywordsAsync
@@ -103,28 +102,6 @@ class BrowserManagementKeywords(LibraryComponent):
     def reload_page(self):
         """Reload the current page"""
         self.loop.run_until_complete(self.get_async_keyword_group().reload_page())
-        # async def reload_page_async():
-        #     await self.ctx.get_current_page().reload()
-        # self.loop.run_until_complete(reload_page_async())
-
-    @keyword
-    def set_timeout(self, timeout):
-        """Sets the timeout that is used by various keywords.
-        The value can be given as a number that is considered to be seconds or as a human-readable string like 1 second.
-        The previous value is returned and can be used to restore the original value later if needed.
-        See the Timeout section above for more information.
-
-        Example:
-
-        | ${orig timeout} =	          | Set Timeout	     | 15 seconds |
-        | Open page that loads slowly |	                 |            |
-        | Set Timeout	              | ${orig timeout}	 |            |
-
-        """
-        orig_timeout = self.ctx.timeout
-        self.ctx.timeout = timestr_to_secs(timeout)
-        self.info('Original timeout is ' + str(orig_timeout) + ' seconds')
-        return orig_timeout
 
     @keyword
     def get_window_count(self):
